backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
sympy
Pillow
pytest
httpx
torch
numpy
